# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is API doc of UniqueStudio Recruitment. For more API information, please see https://app.apifox.com/project/2985744",
        "title": "UniqueStudio Recruitment API",
        "contact": {
            "email": "wwbstar07@gmail.com"
        },
        "version": "0.1"
    },
    "host": "https://dev.back.recruitment2023.hustunique.com/",
    "paths": {
...
        "/sms": {
            "post": {
                "description": "Send sms to user, include Accept, Reject, detailed information reference https://uniquestudio.feishu.cn/docx/Yh96d2DoyoCe6zxlR0ecSU5snDd?from=from_copylink",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sms"
                ],
                "summary": "Send sms",
                "operationId": "send_sms",
                "parameters": [
                    {
                        "description": "sms body params",
                        "name": "pkg.SendSMSOpts",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg.SendSMSOpts"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.JSONResult"
                        }
                    },
                    "400": {
                        "description": "code is not 0 and msg not empty",
                        "schema": {
                            "$ref": "#/definitions/common.JSONResult"
                        }
                    }
                }
            }
        },
        "/sms/code": {
            "post": {
                "description": "Send code to admin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sms"
                ],
                "summary": "Send code",
                "operationId": "send_code",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.JSONResult"
                        }
                    },
                    "400": {
                        "description": "code is not 0 and msg not empty",
                        "schema": {
                            "$ref": "#/definitions/common.JSONResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.JSONResult": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "pkg.SMSType": {
            "type": "string",
            "enum": [
                "Accept",
                "Reject"
            ],
            "x-enum-varnames": [
                "Accept",
                "Reject"
            ]
        },
        "pkg.SendSMSOpts": {
            "type": "object",
            "required": [
                "current",
                "next",
                "type"
            ],
            "properties": {
                "aids": {
                    "description": "the applications will be sent sms",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "current": {
                    "description": "the application current step",
                    "allOf": [
                        {
                            "$ref": "#/definitions/pkg.Step"
                        }
                    ]
                },
                "meeting_id": {
                    "type": "string"
                },
                "next": {
                    "description": "the application next step",
                    "allOf": [
                        {
                            "$ref": "#/definitions/pkg.Step"
                        }
                    ]
                },
                "place": {
                    "description": "the next step(interview/test) place",
                    "type": "string"
                },
                "rest": {
                    "type": "string"
                },
                "time": {
                    "description": "the next step(interview/test) time",
                    "type": "string"
                },
                "type": {
                    "description": "the candidate status : Pass or Fail",
                    "allOf": [
                        {
                            "$ref": "#/definitions/pkg.SMSType"
                        }
                    ]
                }
            }
        },
        "pkg.Step": {
            "type": "string",
            "enum": [
                "SignUp",
                "WrittenTest",
                "GroupTimeSelection",
                "GroupInterview",
                "OnlineGroupInterview",
                "StressTest",
                "TeamTimeSelection",
                "TeamInterview",
                "OnlineTeamInterview",
                "Pass"
            ],
            "x-enum-comments": {
                "GroupInterview": "组面",
                "GroupTimeSelection": "组面时间选择",
                "OnlineGroupInterview": "在线组面",
                "OnlineTeamInterview": "在线群面",
                "Pass": "通过",
                "SignUp": "报名",
                "StressTest": "熬测",
                "TeamInterview": "群面",
                "TeamTimeSelection": "面试时间选择",
                "WrittenTest": "笔试"
            },
            "x-enum-varnames": [
                "SignUp",
                "WrittenTest",
                "GroupTimeSelection",
                "GroupInterview",
                "OnlineGroupInterview",
                "StressTest",
                "TeamTimeSelection",
                "TeamInterview",
                "OnlineTeamInterview",
                "Pass"
            ]
        },
...
    },
...
}